+++ insane/constants.py
@@ -1,0 +1,1 @@
+d2r = 3.14159265358979323846264338327950288/180
