+++ insane/converters.py
@@ -1,0 +1,28 @@
+import math
+
+from .constants import d2r
+
+
+# These functions are not to stay here... -TAW
+def vector(v):
+    if type(v) == str and "," in v:
+        return [float(i) for i in v.split(",")]
+    return float(v)
+
+
+def pdbBoxRead(a):
+    # Convert a PDB CRYST1 entry to a lattice definition.
+    # Convert from Angstrom to nanometer
+    fa, fb, fc, aa, ab, ac = [float(i) for i in a.split()[1:7]]
+    ca, cb, cg, sg         = math.cos(d2r*aa), math.cos(d2r*ab), math.cos(d2r*ac) , math.sin(d2r*ac)
+    wx, wy                 = 0.1*fc*cb, 0.1*fc*(ca-cb*cg)/sg
+    wz                     = math.sqrt(0.01*fc*fc - wx*wx - wy*wy)
+    return [0.1*fa, 0, 0, 0.1*fb*cg, 0.1*fb*sg, 0, wx, wy, wz]
+
+
+def box3d(a):
+    x = [ float(i) for i in a.split(",") ] + 6*[0]
+    if len(x) == 12: # PDB format
+        return pdbBoxRead("CRYST1 "+" ".join([str(i) for i in x]))
+    else:            # GRO format
+        return x[0],x[3],x[4],x[5],x[1],x[6],x[7],x[8],x[2]
